Added functions for NTP

+++ docker/ntp/update_ntp_config.sh
@@ -15,13 +15,25 @@ PORT="$2"
 
 # Backup existing configuration
 echo "Backing up current Chrony configuration to $BACKUP_CONF..."
-cp "$CHRONY_CONF" "$BACKUP_CONF"
+
+if [ -f "$CHRONY_CONF" ]; then
+  cp "$CHRONY_CONF" "$BACKUP_CONF"
+else
+  echo "/etc/chrony/chrony.conf does not exist, skipping..."
+fi
 
 # Update configuration
 echo "Updating Chrony configuration with server $SERVER and port $PORT..."
 cat <<EOL > "$CHRONY_CONF"
 # Updated Chrony configuration
 server $SERVER iburst port $PORT
+
+# The rtcsync directive enables a mode where the system time is periodically copied to the RTC
+# Disable this temporarily during major offsets
+# rtcsync
+
+# Allow stepping the system clock for larger offsets
+makestep 86400 10
 
 # Allow NTP client requests from localhost only
 allow 127.0.0.1
@@ -35,11 +47,3 @@ EOL
 
 echo "Chrony configuration updated successfully."
 
-# Restart Chrony to apply new configuration
-systemctl restart chronyd
-if [ $? -eq 0 ]; then
-    echo "Chrony restarted successfully with the new configuration."
-else
-    echo "Failed to restart Chrony. Check your configuration."
-    exit 1
-fi
